refactor(galgame): split screen_classifier.py into 5 sub-modules
按 galgame-pr4-split-plan.md 第 4 部分纯文件移动 + re-export:
- _stage_keywords.py: 9 个关键词常量 (_TITLE_KEYWORDS .. _GAME_OVER_KEYWORDS)
- _ocr_utils.py: 共享叶子工具 (_cluster_count, _visible_len, _bounded_raw_text,
  _dedupe_preserve_order, _bounded_debug_value, _float, _confidence) + 行长/对话正则常量
- _layout.py: 布局特征 (_box_outside_unit_space, _layout_features, _clamp_unit,
  _nondegenerate_unit_interval, _normalized_bounds, _coerce_rect)
- _templates.py: 模板叶子函数 (_template_regions, _template_region_hits,
  _template_string_list, _template_matches_context, _template_context_score)
- _ocr_pipeline.py: OCR 流水线 + 文本处理 + 各 _looks_like_* 匹配器
- screen_classifier.py 保留: ScreenClassification/_OcrRegion 数据类、
  classify_screen_from_ocr/analyze_screen_visual_features/
  classify_screen_awareness_model/normalize_screen_type 公开函数、
  _classification_from_templates/_classification_from_visual/_classified 分类器

循环依赖处理 (与 Step 2 一致的延迟 import 模式):
- _ocr_utils._dedupe_preserve_order 延迟 import _normalize_for_match
- _ocr_pipeline._coerce_ocr_regions 延迟 import _OcrRegion (运行时构造)

外部公开符号 (classify_screen_from_ocr/normalize_screen_type/
analyze_screen_visual_features/classify_screen_awareness_model/
ScreenClassification/SCREEN_UI_ELEMENT_LIMIT) 通过 re-export 保持原 import 路径。
