+++ plugin/plugins/galgame_plugin/_stage_keywords.py
@@ -1,0 +1,167 @@
+from __future__ import annotations
+
+
+_TITLE_KEYWORDS = (
+    "start",
+    "new game",
+    "newgame",
+    "continue",
+    "load game",
+    "extra",
+    "gallery",
+    "start game",
+    "开始",
+    "開始",
+    "新游戏",
+    "新遊戲",
+    "继续",
+    "繼續",
+    "载入",
+    "載入",
+    "はじめから",
+    "つづきから",
+    "スタート",
+    "ロード",
+    "タイトル",
+    "コンティニュー",
+)
+_TITLE_EXIT_KEYWORDS = ("quit", "exit", "结束", "終了", "退出")
+_SAVE_LOAD_KEYWORDS = (
+    "save",
+    "load",
+    "quick save",
+    "quick load",
+    "autosave",
+    "auto save",
+    "slot",
+    "page",
+    "保存",
+    "存档",
+    "存檔",
+    "读档",
+    "讀檔",
+    "载入",
+    "載入",
+    "セーブ",
+    "ロード",
+    "スロット",
+    "ページ",
+)
+_CONFIG_KEYWORDS = (
+    "config",
+    "configuration",
+    "settings",
+    "setting",
+    "volume",
+    "sound",
+    "voice",
+    "bgm",
+    "se",
+    "text speed",
+    "message speed",
+    "window mode",
+    "fullscreen",
+    "设置",
+    "設定",
+    "音量",
+    "声音",
+    "語音",
+    "语音",
+    "文字速度",
+    "对话速度",
+    "對話速度",
+    "全屏",
+    "画面",
+    "畫面",
+    "コンフィグ",
+    "設定",
+    "オプション",
+    "音量",
+    "ボイス",
+    "メッセージ",
+)
+_BACK_KEYWORDS = ("back", "return", "close", "戻る", "返回", "取消")
+_BACKLOG_KEYWORDS = (
+    "backlog",
+    "message log",
+    "dialogue log",
+    "dialog log",
+    "text log",
+    "history",
+    "履歴",
+    "会話履歴",
+    "バックログ",
+    "ログ",
+    "历史",
+    "歷史",
+    "历史记录",
+    "歷史記錄",
+    "对话历史",
+    "對話歷史",
+    "对白历史",
+    "對白歷史",
+    "文本历史",
+    "文本歷史",
+    "讯息记录",
+    "訊息記錄",
+)
+_GALLERY_KEYWORDS = (
+    "gallery",
+    "cg mode",
+    "cg",
+    "scene replay",
+    "replay",
+    "album",
+    "memories",
+    "extra",
+    "回想",
+    "鉴赏",
+    "鑑賞",
+    "画廊",
+    "圖鑑",
+    "图鉴",
+    "相册",
+    "ギャラリー",
+    "シーン回想",
+    "鑑賞モード",
+)
+_MINIGAME_KEYWORDS = (
+    "minigame",
+    "mini game",
+    "score",
+    "combo",
+    "time limit",
+    "remaining time",
+    "hp",
+    "操作説明",
+    "スコア",
+    "コンボ",
+    "残り時間",
+    "小游戏",
+    "小遊戲",
+    "得分",
+    "连击",
+    "連擊",
+    "剩余时间",
+    "剩餘時間",
+)
+_GAME_OVER_KEYWORDS = (
+    "game over",
+    "bad end",
+    "dead end",
+    "retry",
+    "try again",
+    "continue?",
+    "return to title",
+    "back to title",
+    "游戏结束",
+    "遊戲結束",
+    "重试",
+    "重試",
+    "再试一次",
+    "再試一次",
+    "返回标题",
+    "返回標題",
+    "タイトルへ",
+    "リトライ",
+)
